double -> float + physics finished

+++ character.h
@@ -16,7 +16,7 @@ using namespace std;
 class Character: public MovableRectangle
 {
 public:
-    Character(Vec2 pos_, double time_betwwen_frames_, double mass, double max_hp_);
+    Character(Vec2 pos_, float time_betwwen_frames_, float mass, float max_hp_);
     
     void loadSpriteFrames(const QString &basePath);
     void load_jump_frames(const QString &basePath);
@@ -34,16 +34,16 @@ public:
     void update_sword_attack();
 
     void draw(QPainter &painter);    
-    void set_x(double x_) {pos.x = x_;}
-    void set_y(double y_) {pos.y = y_;}
-    void set_x_expected(double x_expected_) { pos_exp.x = x_expected_ ; }
-    void set_y_expected(double y_expected_) { pos_exp.y = y_expected_; }
-    void update_vel(double dt){ v.x = (get_x_expected() - get_x()) / dt; v.y = (get_y_expected() - get_y()) / dt; }
-    void update_expected_pos(double dt){ pos_exp.x = get_x() + v.x * dt; pos_exp.y = get_y() + v.y * dt; }
+    void set_x(float x_) {pos.x = x_;}
+    void set_y(float y_) {pos.y = y_;}
+    void set_x_expected(float x_expected_) { pos_exp.x = x_expected_ ; }
+    void set_y_expected(float y_expected_) { pos_exp.y = y_expected_; }
+    void update_vel(float dt){ v.x = (get_x_expected() - get_x()) / dt; v.y = (get_y_expected() - get_y()) / dt; }
+    void update_expected_pos(float dt){ pos_exp.x = get_x() + v.x * dt; pos_exp.y = get_y() + v.y * dt; }
     void update_pos(){ pos.x = pos_exp.x; pos.y = pos_exp.y; }
-    void update_expected_pos_collision(double delta_x, double delta_y){ pos_exp.x += delta_x; pos_exp.y += delta_y; }
-    void set_c_scale(double s) {c_scale = s;}
-    void set_hitbox(QPixmap& sprite, vector<vector<double>>* asset_dims_, vector<vector<double>>* character_dims_, vector<vector<double>>* sword_dims_);
+    void update_expected_pos_collision(float delta_x, float delta_y){ pos_exp.x += delta_x; pos_exp.y += delta_y; }
+    void set_c_scale(float s) {c_scale = s;}
+    void set_hitbox(QPixmap& sprite, vector<vector<float>>* asset_dims_, vector<vector<float>>* character_dims_, vector<vector<float>>* sword_dims_);
     void set_draw_baxes(bool d) { draw_baxes = d; }
     void set_right(bool r);
     void set_right_pressed(bool r) {right_pressed = r;}
@@ -53,34 +53,34 @@ public:
     void set_sliding(bool s);
     void set_lowering(bool s);
     void set_sword_attacking(bool a);
-    void set_speed_move(double s) {speed_move = s;}
-    void set_speed_jump(double s) {speed_jump = s;}
-    void set_speed_run(double s) {speed_run = s;}
-    void set_hp(double hp_) {hp = hp_; lifebar->set_percentage(hp/max_hp);}
+    void set_speed_move(float s) {speed_move = s;}
+    void set_speed_jump(float s) {speed_jump = s;}
+    void set_speed_run(float s) {speed_run = s;}
+    void set_hp(float hp_) {hp = hp_; lifebar->set_percentage(hp/max_hp);}
     void set_first_hit_sword_attack(bool f_h) {first_hit_sword_attack = f_h;}
-    void set_lifebar_dims(double x, double y, double w, double h);
+    void set_lifebar_dims(float x, float y, float w, float h);
 
-    double get_rest() const override { return 0.0; } // No bounce for characters
+    float get_rest() const override { return 0.0; } // No bounce for characters
 
     // Drawing positions (with sprite offset)
-    double get_y() const { return pos.y;} 
-    double get_x() const { return pos.x;} 
-    double get_x_expected() const { return pos_exp.x;} 
-    double get_y_expected() const { return pos_exp.y;}
+    float get_y() const { return pos.y;} 
+    float get_x() const { return pos.x;} 
+    float get_x_expected() const { return pos_exp.x;} 
+    float get_y_expected() const { return pos_exp.y;}
 
 
     void set_moving(bool m);
 
-    double get_mass() const { return mass; } 
-    double get_x_sliding(double t);
-    double get_x_sword_attacking(double t);
-    double get_c_scale() const { return c_scale; }
-    double get_hp() const { return hp; }
-    double get_current_sprite_width() {return currentSprite->width();}
-    double get_sword_attack_damages() { return sword_attack_damages; }
-    vector<double> get_current_asset_dims() {return current_asset_dims; }
-    vector<double> get_current_character_dims() {return current_character_dims; }
-    vector<double> get_current_sword_dims() {return current_sword_dims; }
+    float get_mass() const { return mass; } 
+    float get_x_sliding(float t);
+    float get_x_sword_attacking(float t);
+    float get_c_scale() const { return c_scale; }
+    float get_hp() const { return hp; }
+    float get_current_sprite_width() {return currentSprite->width();}
+    float get_sword_attack_damages() { return sword_attack_damages; }
+    vector<float> get_current_asset_dims() {return current_asset_dims; }
+    vector<float> get_current_character_dims() {return current_character_dims; }
+    vector<float> get_current_sword_dims() {return current_sword_dims; }
 
 
 
@@ -100,8 +100,8 @@ public:
     void checkBounds(int windowWidth);   
 
 private:
-    double x;
-    double y;
+    float x;
+    float y;
     bool moving = false;
     bool jumping = false;
     bool sliding = false;
@@ -117,27 +117,27 @@ private:
     
     // Sprite animation
     QVector<QPixmap> move_frames;
-    vector<vector<double>> move_frames_asset_dims;
-    vector<vector<double>> jump_frames_asset_dims;
-    vector<vector<double>> lower_frames_asset_dims;
-    vector<vector<double>> sword_attack_frames_asset_dims;
-    vector<vector<double>> move_frames_character_dims;
-    vector<vector<double>> jump_frames_character_dims;
-    vector<vector<double>> lower_frames_character_dims;
-    vector<vector<double>> sword_attack_frames_character_dims;
-    vector<vector<double>> sword_attack_frames_sword_dims;
+    vector<vector<float>> move_frames_asset_dims;
+    vector<vector<float>> jump_frames_asset_dims;
+    vector<vector<float>> lower_frames_asset_dims;
+    vector<vector<float>> sword_attack_frames_asset_dims;
+    vector<vector<float>> move_frames_character_dims;
+    vector<vector<float>> jump_frames_character_dims;
+    vector<vector<float>> lower_frames_character_dims;
+    vector<vector<float>> sword_attack_frames_character_dims;
+    vector<vector<float>> sword_attack_frames_sword_dims;
     QVector<QPixmap> jump_frames;
     QVector<QPixmap> slide_frames;
     QVector<QPixmap> lower_frames;
     QVector<QPixmap> sword_attack_frames;
-    QVector<double> bounds;
+    QVector<float> bounds;
     Lifebar *lifebar;
     QPixmap *currentSprite = nullptr;
-    vector<double> current_asset_dims;
-    vector<double> current_character_dims;
-    vector<double> current_sword_dims;
-    vector<double> idle_dims;
-    vector<double> idle_character_dims;
+    vector<float> current_asset_dims;
+    vector<float> current_character_dims;
+    vector<float> current_sword_dims;
+    vector<float> idle_dims;
+    vector<float> idle_character_dims;
     
 
     QPixmap idleFrame;
@@ -153,31 +153,31 @@ private:
     std::string slide_dir;
     std::string sword_attack_dir = "right";
     
-    double mass {2.0};
-    double max_hp {100.0};
-    double jump_height {50.0};
-    double speed_move;
-    double speed_run;
-    double speed_jump {2500.0};  // Initial upward velocity for jumps
-    double slide_dist {4.0};
-    double sword_attack_dist {1.0};
-    double total_jump_time {0.5};
-    double total_slide_time {0.5};
-    double total_sword_attack_time {0.5};
-    double time_between_slides {1.0};
-    double time_between_sword_attacks {1.0};
-    double lower_time {1.0};
-    double jump_time;
-    double slide_time;
-    double sword_attack_time;
-    double walk_step_time {10000.0};
-    double total_walk_step_time {0.5};  // Total duration for one walk cycle
-    double time_between_frames;
-    double c_scale = 100.0;
+    float mass {2.0};
+    float max_hp {100.0};
+    float jump_height {50.0};
+    float speed_move;
+    float speed_run;
+    float speed_jump {2500.0};  // Initial upward velocity for jumps
+    float slide_dist {4.0};
+    float sword_attack_dist {1.0};
+    float total_jump_time {0.5};
+    float total_slide_time {0.5};
+    float total_sword_attack_time {0.5};
+    float time_between_slides {1.0};
+    float time_between_sword_attacks {1.0};
+    float lower_time {1.0};
+    float jump_time;
+    float slide_time;
+    float sword_attack_time;
+    float walk_step_time {10000.0};
+    float total_walk_step_time {0.5};  // Total duration for one walk cycle
+    float time_between_frames;
+    float c_scale = 100.0;
     bool draw_baxes = true;
-    double base_y {0.0};  // Base Y position when jump starts
-    double hp {100.0};
-    double sword_attack_damages {10.0};
+    float base_y {0.0};  // Base Y position when jump starts
+    float hp {100.0};
+    float sword_attack_damages {10.0};
 };
 
 #endif // CHARACTER_H
